do the pathfinding
it does not always find the best path yet, there are at least two problems i see:

The obvious one is that this only uses vertices of the nav mesh as waypoints (except start & end). 
So two points sitting on two neighboring triangles will always travel over one of the vertices, which is rarely the shortest path.
Funnelling seems to be the correct way to solve this, TBD. 

But there also seems to be cases where our implementation fails to find the shortest route over these edges.
I think the check if the current node is one of the edges of our target vertice is to early. 
If we find one of these three points as our current looked at node we immediately finish, but just because that points belongs
to the same triangle as our goal does not mean that there is no shorter path using more points to get to another vertix of this triangle
So we must change that check and tread these points as normal nodes, + handle the special case that our goal is a neighbor of these three.

Other than that: it works, but needs a lot of cleanup!